POC for working with submodules (it's possible!)
For now it works like that: if the path to the file status of which we want to
get contains 'Plugins' string, then search this file in submodule by running
'git submodule foreach git status <usual arguments for git status>'.
(So if your submodule is not under 'Plugins' directory, it won't work)

Also getting the array of paths to submodules - for later. The plan is to check the files
against these paths and if the file is in one of a paths, pass this path
as a root git directory. This way, running the same 'git status' command
will give us a status of a submodule's file while runnig it on a default
root would fail.